requirements.txt
# RedTeam Steganography Toolkit Dependencies

# GUI Framework
PyQt5>=5.15.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# Cryptography
cryptography>=3.4.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
